Revert "#163787456 Write tests for all functionality"
