--- .travis.yml
@@ -1,18 +1,0 @@
-language: node_js
-node_js:
-  - "10"
-addons: code_climate
-env:
-  global:
-    - CC_TEST_REPORTER_ID=b6077c510544b4b4e54db7d362ee50accf90da669ba7a0a42e99ccc83138d805
-cache:
-  directories:
-    - node_modules
-before_script:
-  - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
-  - chmod +x ./cc-test-reporter
-  - ./cc-test-reporter before-build
-script:
-  - yarn test --detectOpenHandles -u
-after_script:
-  - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
