+++ package.json
@@ -9,8 +9,7 @@
   "private": false,
   "scripts": {
     "start": "cross-env node server",
-    "webpack": "webpack -wd",
-    "test": "jest --coverage -u"
+    "webpack": "webpack -wd"
   },
   "devDependencies": {
     "babel-eslint": "^10.0.1",
@@ -26,31 +25,26 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "axios": "^0.18.0",
-    "axios-mock-adapter": "^1.16.0",
     "babel-loader": "^8.0.5",
     "babel-regenerator-runtime": "^6.5.0",
     "bootstrap": "^4.2.1",
     "classnames": "^2.2.6",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.0",
-    "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.9.1",
     "express": "^4.16.4",
     "express-yields": "^1.1.2",
     "file-loader": "^3.0.1",
     "fs-extra": "^7.0.1",
     "gitignore": "^0.6.0",
     "html-webpack-plugin": "^3.2.0",
-    "jest": "^24.1.0",
-    "jest-webpack": "^0.5.1",
     "materialize-css": "^1.0.0",
     "moment": "^2.24.0",
+    "msg-notify": "^1.1.3",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
     "redux": "^4.0.1",
-    "redux-mock-store": "^1.5.3",
     "redux-thunk": "^2.3.0",
     "regenerator-runtime": "^0.13.1",
     "style-loader": "^0.23.1",
