+++ src/components/common/FoodList.js
@@ -8,10 +8,15 @@ class  FoodList extends React.Component {
       let elems = document.querySelectorAll('.collapsible');
       M.Collapsible.init(elems);
     });
+    document.addEventListener('mouseUp', () => {
+      let elems = document.querySelectorAll('.tooltipped');
+      M.Tooltip.init(elems);
+    });
   }
 
   render() {
     const {foodItems} = this.props;
+    console.log(foodItems);
     return (
       <ul className="collapsible popout">
         {foodItems.map(foodItem => <FoodItem foodItem={foodItem} />)}
